SAT solvers - remove warnings - rollback

+++ Hydra/external/satSolvers/MiniSat/core/Solver.cc
@@ -504,7 +504,7 @@ struct reduceDB_lt {
 };
 void Solver::reduceDB() {
     int i, j;
-    float extra_lim = (float)(cla_inc / learnts.size());   // Remove any clause below this activity
+    double extra_lim = cla_inc / learnts.size();   // Remove any clause below this activity
 
     sort(learnts, reduceDB_lt(ca));
     // Don't delete binary or locked clauses. From the rest, delete clauses from the first half
